Version bump
Avoid running deborphan for Salted boxes.

+++ script/minimize.sh
@@ -32,12 +32,14 @@ apt-get -y autoremove --purge
 apt-get -y autoclean
 apt-get -y clean
 
-# Clean up orphaned packages with deborphan
-apt-get -y install deborphan
-while [ -n "$(deborphan --guess-all --libdevel)" ]; do
-    deborphan --guess-all --libdevel | xargs apt-get -y purge
-done
-apt-get -y purge deborphan dialog
+if [ "$CM" != "salt" ]; then
+    # Clean up orphaned packages with deborphan
+    apt-get -y install deborphan
+    while [ -n "$(deborphan --guess-all --libdevel)" ]; do
+        deborphan --guess-all --libdevel | xargs apt-get -y purge
+    done
+    apt-get -y purge deborphan dialog
+fi
 
 echo "==> Removing man pages"
 rm -rf /usr/share/man/*
